scraper/requirements.txt
requests==2.28.*
google-cloud-storage==2.4.*